backend/requirements.txt
Flask==2.1.2
PyJWT==2.6.0
cryptography==3.4.8
requests==2.26.0
